Moved AddListInput from ListShoppingLists
