+++ src/components/Common/FormControl/AddListInput.less
@@ -1,0 +1,20 @@
+.itemWrapper {
+    margin: 20px 0;
+    border: 1px solid #ebebeb;
+    border-radius: 10px;
+    padding: 20px 20px;
+    box-sizing: border-box;
+    font-size: 18pt;
+    background-color: #ffffff;
+}
+
+.CustomInput {
+    border: 0;
+    border-radius: 2px;
+    width: 100%;
+    outline: none;
+    -webkit-appearance: none;
+    box-shadow: 0 0 0 2px #015ccb;
+    font-size: 18pt;
+    font-family: "Lab Grotesque", serif;
+}
