+++ src/components/ListShoppingLists/ListShoppingLists.less
@@ -32,14 +32,3 @@
     box-sizing: border-box;
     width: 100%;
 }
-
-.CustomInput {
-    border: 0;
-    border-radius: 2px;
-    width: 100%;
-    outline: none;
-    -webkit-appearance: none;
-    box-shadow: 0 0 0 2px #015ccb;
-    font-size: 18pt;
-    font-family: "Lab Grotesque", serif;
-}
